+++ tests/rx/write_byte.cpp
@@ -3,19 +3,19 @@
 TEST_F(RxTest, write_byte_operations_mode) {
   // Don't write any CV which might trigger config (e.g. 1, 28, ...)!
   auto cv_addr{RandomInterval(30u, smath::pow(2u, 10u) - 1u)};
-  auto cv_value{RandomInterval<uint8_t>(0u, 255u)};
+  auto byte{RandomInterval<uint8_t>(0u, 255u)};
 
-  EXPECT_CALL(_mock, writeCv(cv_addr, cv_value));
-  ReceiveAndExecuteTwoIdenticalCvWritePackets(cv_addr, cv_value);
+  EXPECT_CALL(_mock, writeCv(cv_addr, byte));
+  ReceiveAndExecuteTwoIdenticalCvWritePackets(_addrs.primary, cv_addr, byte);
 }
 
 TEST_F(RxTest, write_byte_operations_mode_requires_two_identical_packets) {
   auto cv_addr{RandomInterval(30u, smath::pow(2u, 10u) - 1u)};
-  auto cv_value{RandomInterval<uint8_t>(0u, 255u)};
+  auto byte{RandomInterval<uint8_t>(0u, 255u)};
   auto packet{
-    dcc::make_cv_access_long_write_packet(_addrs.primary, cv_addr, cv_value)};
+    dcc::make_cv_access_long_write_packet(_addrs.primary, cv_addr, byte)};
 
-  EXPECT_CALL(_mock, writeCv(cv_addr, cv_value)).Times(0);
+  EXPECT_CALL(_mock, writeCv(cv_addr, byte)).Times(0);
   Receive(packet);
   Execute();
 }
@@ -25,11 +25,11 @@ TEST_F(RxTest, write_byte_service_mode) {
 
   // Don't write any CV which might trigger config (e.g. 1, 28, ...)!
   auto cv_addr{RandomInterval(30u, smath::pow(2u, 10u) - 1u)};
-  auto cv_value{RandomInterval<uint8_t>(0u, 255u)};
-  auto packet{dcc::make_cv_access_long_write_service_packet(cv_addr, cv_value)};
+  auto byte{RandomInterval<uint8_t>(0u, 255u)};
+  auto packet{dcc::make_cv_access_long_write_service_packet(cv_addr, byte)};
 
   // 5 or more identical packets
-  EXPECT_CALL(_mock, writeCv(cv_addr, cv_value));
+  EXPECT_CALL(_mock, writeCv(cv_addr, byte));
   for (auto i{0uz}; i < 5uz; ++i) {
     Receive(packet);
     Execute();